# services/friend/api/openapi/api.yaml
openapi: 3.0.0
info:
  title: Friend Service API
  version: 1.0.0
  description: API for managing friends and friend requests
paths:
  /add-friend:
    post:
      summary: Send a friend request
      requestBody:
        description: Friend request details
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FriendRequest'
        required: true
      responses:
        '200':
          description: Friend request sent successfully
        '400':
          description: Invalid request payload
        '500':
          description: Internal server error

  /friends:
    get:
      summary: Retrieve the friend list for the authenticated user
      parameters:
        - name: userId
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: List of friends
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Friend'
        '400':
          description: Invalid request parameters
        '500':
          description: Internal server error

  /live:
    get:
      summary: Health check for service liveness
      responses:
        '200':
          description: Service is alive
        '500':
          description: Internal server error

  /ready:
    get:
      summary: Health check for service readiness
      responses:
        '200':
          description: Service is ready
        '500':
          description: Internal server error

  /respond-friend-request:
    post:
      summary: Respond to a friend request
      requestBody:
        description: Friend request response details
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FriendRequestResponse'
        required: true
      responses:
        '200':
          description: Response to friend request processed successfully
        '400':
          description: Invalid request payload
        '500':
          description: Internal server error

components:
  schemas:
    FriendRequest:
      type: object
      properties:
        userId:
          type: string
          description: ID of the user sending the friend request
        friendId:
          type: string
          description: ID of the friend to add
      required:
        - userId
        - friendId

    Friend:
      type: object
      properties:
        userId:
          type: string
          description: ID of the user
        friendId:
          type: string
          description: ID of the friend
      required:
        - userId
        - friendId

    FriendRequestResponse:
      type: object
      properties:
        friendId:
          type: string
          description: ID of the friend
        response:
          type: string
          enum:
            - accepted
            - rejected
          description: Response to the friend request
      required:
        - friendId
        - response